+++ src/main/java/io/github/ivanmilovanov/agentic/cli/runner/api/AgentRunner.java
@@ -8,8 +8,8 @@ import io.github.ivanmilovanov.agentic.cli.runner.model.AgentResultDto;
 public interface AgentRunner {
 
     /** Запускает агента с произвольным пользовательским промптом. */
-    AgentResultDto executeUserPrompt(String prompt) throws Exception;
+    AgentResultDto execute(String prompt) throws Exception;
 
     /** Запускает агента с указанным скилом и промптом для него. */
-    AgentResultDto executeSkillPrompt(String skillName, String prompt) throws Exception;
+    AgentResultDto executeSkill(String skillName, String prompt) throws Exception;
 }